Moved utils.js
